requirements.txt
streamlit==1.22.0
pillow==9.5.0
diffusers==0.16.1
transformers==4.29.2
opencv-python-headless==4.7.0.72
numpy==1.23.5
safetensors==0.3.1
controlnet_aux==0.0.6
pytorch_lightning==1.9.0
torch==1.10.0
torchdiffeq==0.2.2
torchsde==0.2.7
matplotlib==3.5.1
